payment_requirements.txt
python-telegram-bot==20.7
flask==3.0.0
python-dotenv==1.0.0
requests==2.32.5